requirements-gpu.txt
cycler==0.10.0
numpy
stt==1.0.0
tensorflow-gpu==1.15
deepspeech-gpu==0.9.3
joblib==0.16.0
kiwisolver==1.2.0
pydub==0.23.1
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn
scipy==1.4.1
six==1.15.0
tqdm==4.44.1
